iotjs: Relocate TizenRT's kconfig out of iotjs subdir
Relocating it outside it will make the module easy to update,
by just replacing iotjs subtree with upstream tree.

For developer conveinance and to ease upstreaming works
(tracking changes, logs etc) the folder can be replaced by submodule,
like other deps in Tizen:RT.
But at the cost of lossing all changes occured in after version import.

For the record, I tried to upstream the Kconfig.runtime file:
https://github.com/Samsung/iotjs/pull/1452

But it wasn't merged, this is making sense since it's only for downstream.

So I am suggesting to avoid mixing upstream and downstream sources,
and keep iotjs as aligned to upstream as possible,
and make rebasing easier to avoid duplication of maintenance.

Change-Id: If4a7d71d862745173f38d009924f3e4ce98cb6d0
Bug: https://github.com/Samsung/TizenRT/pull/1277
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>